backend/requirements.txt
annotated-types==0.5.0
asgiref==3.7.2
atomicwrites==1.4.1
attrs==19.3.0
beautifulsoup4==4.11.1
certifi==2023.7.22
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.2.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
coverage==7.1.0
cryptography==41.0.2
defusedxml==0.7.1
Django==3.2.15
django-debug-toolbar==2.2
django-filter==21.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-extra-fields==3.5.0
Faker==12.0.1
filetype==1.2.0
flake8==6.0.0
idna==3.4
importlib-metadata==1.5.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mccabe==0.7.0
mixer==7.1.2
more-itertools==8.2.0
oauthlib==3.2.2
packaging==20.1
Pillow==10.0.0
pluggy==0.13.1
psycopg2==2.9.6
py==1.8.1
pycodestyle==2.10.0
pycparser==2.21
pydantic==2.0.3
pydantic_core==2.3.0
pyflakes==3.0.1
pygismeteo==5.2.0
pygismeteo-base==4.4.2
PyJWT==2.8.0
pyparsing==2.4.6
pytest==5.3.5
pytest-django==3.8.0
pytest-pythonpath==0.7.3
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.2
sorl-thumbnail==12.6.3
soupsieve==2.3.2.post1
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.4
wcwidth==0.1.8
zipp==2.2.0
